Create a script for the cards that will be played.
* StandardCarts class: this cards are the numeric cards, from 0 to 9.
Every card has a color and a content.

* generate_carts function: this function make an array of a pair of
numbers from 0 to 9, take this array an pass it for the shuffle cards
function to mix it, then create another array of colors that gonna be
the colors from each cart, create a new array and fill it from the
creation of a object for every number with his color, and finally the
array of objects is returned.

* shuffle_cards function: this function receive an array and mix their
values with a variation of the Fisher-Yates algorithm.

The other cards class are already created, but his functionality is
still in process.